Sweet alert

+++ src/api/products.js
@@ -63,11 +63,24 @@ export function updateProduct(id, payload) {
   });
 }
 
-export function deleteProduct({ id }) {
-  return http.delete(`/products/${id}`).then((response) => {
-    const { data: json } = response;
-    return {
-      data: transformProduct(json.data),
-    };
-  });
+// export function deleteProduct({ id }) {
+//   return http.delete(`/products/${id}`).then((response) => {
+//     const { data: json } = response;
+//     return {
+//       data: transformProduct(json.data),
+//     };
+//   });
+// }
+
+export function deleteProduct({ id }, confirm = true) {
+  if (confirm) {
+    return http.delete(`/products/${id}`).then((response) => {
+      const { data: json } = response;
+      return {
+        data: transformProduct(json.data),
+      };
+    });
+  } else {
+    return Promise.resolve({ data: null });
+  }
 }
